refactor: router 이름 변경 및 주석 추가
- homeRouter -> indexRouter
- app.ts 미들웨어 주석 추가
